+++ model_description/models/dvigatel.js
@@ -11,17 +11,17 @@ module.exports = {
         },
         Rabochij_obyom_dvigatelya: {
             label: "Рабочий объем двигателя", value: "", units: "л",
-            type: 'INTEGER',
+            type: 'FLOAT',
             units_gloabal_id: "273472400",
         },
         Moshchnost: {
             label: "Мощность", value: "", units: "кВт",
-            type: 'INTEGER',
+            type: 'FLOAT',
             units_gloabal_id: "273472448",
         },
         Krutyashch_moment: {
             label: "Крутящий момент", value: "", units: "Н/м",
-            type: 'INTEGER',
+            type: 'FLOAT',
             units_gloabal_id: "273472508",
         },
         cilindry_id: {
